PhrasIQ/requirements.txt
pandas
python-dotenv
langchain>=0.1.0
langchain-openai
langchain-community
faiss-cpu
streamlit
openai
